Moving moongoose configurations to config file
